move vho query service to vho module
