+++ VideoWeb/VideoWeb/ClientApp/src/app/vh-officer/vh-officer.module.ts
@@ -27,7 +27,7 @@ import { VhOfficerRoutingModule } from './vh-officer-routing.module';
 import { VhoChatComponent } from './vho-chat/vho-chat.component';
 import { VhoHearingListComponent } from './vho-hearing-list/vho-hearing-list.component';
 import { VhoHearingsFilterComponent } from './vho-herings-filter/vho-hearings-filter.component';
-import { VhoQueryService } from '../services/vho-query-service.service';
+import { VhoQueryService } from './services/vho-query-service.service';
 
 @NgModule({
     declarations: [
